Bugs gefixt
optinal.isEmpty abfragen, equals bei Stringvergleich nutzen

+++ src/main/java/de/hsrm/mi/swtpro/pflamoehus/order/orderdetails/orderdetailsservice/OrderDetailsServiceImpl.java
@@ -120,7 +120,10 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
     @Override
     public void deleteOrderDetail(long id) {
         Optional<OrderDetails> od = findByID(id);
-        orderDetailsRepo.delete(od.get());
+        if(od.isPresent()){
+            orderDetailsRepo.delete(od.get());
+        }
+        
     }
 
 }
